--- css/graph.css
@@ -1,119 +1,0 @@
-/* fuente de google fonts */
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap");
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-}
-
-body {
-  margin: 0;
-  font-family: "Roboto", sans-serif;
-  background-color: #121414;
-  font-weight: bold;
-  color: #ffffff;
-  overflow: hidden; /* Para ocultar el desbordamiento de la barra de desplazamiento */
-}
-
-#graph {
-  height: calc(
-    100vh - 100px
-  ); /* Calcular la altura del gráfico para ocupar todo el espacio disponible menos el área del botón */
-  overflow: auto; /* Agregar desplazamiento */
-}
-
-h1 {
-  font-size: 35px;
-  text-align: center;
-  margin-top: 20px;
-  font-weight: bold;
-}
-h2 {
-  font-size: 20px;
-  text-align: center;
-  margin-top: 5px;
-}
-
-.node {
-  fill: #ff98e5; /* Color azul */
-  stroke: #121414;
-  stroke-width: 2px;
-  cursor: move;
-  stroke-opacity: 0.5;
-}
-
-@keyframes animateDash {
-  to {
-    stroke-dashoffset: -10; /* Ajusta el valor según el tamaño de tu línea */
-  }
-}
-
-.link {
-  stroke: #ffffff; /* Color rojo */
-  stroke-width: 3px;
-  stroke-opacity: 0.7;
-  stroke-dasharray: 5, 5;
-  stroke-linecap: round;
-  animation: animateDash 0.7s linear infinite; /* Aplica la animación infinita */
-}
-
-.link-label {
-  font-size: 20px;
-  fill: #ffffff; /* Color del texto */
-}
-
-#container {
-  text-align: center;
-  padding: 10px;
-  background-color: #000000;
-}
-
-input[type="file"] {
-  display: none;
-}
-
-div {
-  padding: 0px;
-}
-
-label {
-  display: inline-block;
-  background-color: #ff98e5;
-  color: #121414;
-  padding: 10px 20px;
-  cursor: pointer;
-  border-radius: 5px;
-  font-size: 16px;
-}
-
-button {
-  display: inline-block;
-  background-color: #1d1d1d;
-  color: #ffffff;
-  padding: 5px 5px;
-  cursor: pointer;
-  border-radius: 5px;
-  font-size: 17px;
-  border: none;
-}
-
-/* Estilo para la barra de desplazamiento en navegadores WebKit */
-*::-webkit-scrollbar {
-  width: 10px; /* Ancho de la barra de desplazamiento */
-}
-
-/* Estilo para el riel de la barra de desplazamiento */
-*::-webkit-scrollbar-track {
-  background: #121212; /* Color de fondo del riel */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento */
-*::-webkit-scrollbar-thumb {
-  background: #555; /* Color del pulgar */
-  border-radius: 5px; /* Radio de borde del pulgar */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento cuando se está desplazando */
-*::-webkit-scrollbar-thumb:hover {
-  background: #ff98e5; /* Color del pulgar al pasar el ratón sobre él */
-}
