--- css/preview.css
@@ -1,144 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Roboto:wght@300&display=swap");
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-}
-
-body {
-  margin: 0;
-  font-family: "Roboto", sans-serif;
-  background-color: #000000;
-  font-weight: bold;
-  color: #ffffff;
-  overflow-y: scroll;
-}
-
-#container {
-  text-align: center;
-  padding: 20px;
-}
-h1 {
-  font-size: 40px;
-  text-align: center;
-  margin-top: 20px;
-}
-
-#file-info {
-  margin-bottom: 20px;
-}
-
-#file-info p {
-  text-align: center;
-  margin: 0;
-  font-size: 18px;
-  color: #ff98e5;
-}
-
-#file-name {
-  font-weight: bold;
-}
-
-#file-content1 {
-  background-color: #121414;
-  padding: 20px;
-  border-radius: 5px;
-  font-size: 14px;
-  color: #ffffff;
-  margin: 0 100px;
-  white-space: pre-wrap;
-  overflow-y: scroll; /* Habilitar el scroll vertical cuando sea necesario */
-  max-height: 80vh;
-}
-
-#automate {
-  background-color: #121414;
-  padding: 20px;
-  border-radius: 5px;
-  font-size: 14px;
-  color: #ffffff;
-  margin: 0 100px;
-  white-space: pre-wrap;
-  overflow-y: scroll; /* Habilitar el scroll vertical cuando sea necesario */
-  max-height: 80vh;
-  margin-bottom: 100px;
-}
-
-#analyze {
-  background-color: #ff98e5;
-  color: #000000;
-  font-size: 16px;
-  font-weight: bold;
-  padding: 10px 20px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-/* Estilo para la barra de desplazamiento en navegadores WebKit */
-#file-content1::-webkit-scrollbar {
-  width: 10px; /* Ancho de la barra de desplazamiento */
-}
-
-/* Estilo para el riel de la barra de desplazamiento */
-#file-content1::-webkit-scrollbar-track {
-  background: #121212; /* Color de fondo del riel */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento */
-#file-content1::-webkit-scrollbar-thumb {
-  background: #555; /* Color del pulgar */
-  border-radius: 5px; /* Radio de borde del pulgar */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento cuando se está desplazando */
-#file-content1::-webkit-scrollbar-thumb:hover {
-  background: #ff98e5; /* Color del pulgar al pasar el ratón sobre él */
-}
-
-/* Estilo para la barra de desplazamiento en navegadores WebKit */
-body::-webkit-scrollbar {
-  width: 10px; /* Ancho de la barra de desplazamiento */
-}
-
-/* Estilo para el riel de la barra de desplazamiento */
-body::-webkit-scrollbar-track {
-  background: #121212; /* Color de fondo del riel */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento */
-body::-webkit-scrollbar-thumb {
-  background: #555; /* Color del pulgar */
-  border-radius: 5px; /* Radio de borde del pulgar */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento cuando se está desplazando */
-body::-webkit-scrollbar-thumb:hover {
-  background: #ff98e5; /* Color del pulgar al pasar el ratón sobre él */
-}
-
-/* Estilo para la barra de desplazamiento en navegadores WebKit */
-#automate::-webkit-scrollbar {
-  width: 10px; /* Ancho de la barra de desplazamiento */
-}
-
-/* Estilo para el riel de la barra de desplazamiento */
-#automate::-webkit-scrollbar-track {
-  background: #121212; /* Color de fondo del riel */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento */
-#automate::-webkit-scrollbar-thumb {
-  background: #555; /* Color del pulgar */
-  border-radius: 5px; /* Radio de borde del pulgar */
-}
-
-/* Estilo para el pulgar de la barra de desplazamiento cuando se está desplazando */
-#automate::-webkit-scrollbar-thumb:hover {
-  background: #ff98e5; /* Color del pulgar al pasar el ratón sobre él */
-}
-
-/* Ignorar por completo y que no se vea el id=file-content */
-#file-content {
-  display: none;
-}
