+++ index.html
@@ -1,47 +1,73 @@
 <!DOCTYPE html>
 <html lang="es">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/css/graph.css" />
-    <title>Grafo Interactivo</title>
+<head>
+    <meta charset="UTF-8">
+    <title>Validador de PDA</title>
+    <!-- Bootstrap CSS -->
+    <link href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" rel="stylesheet">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.3.1/styles/default.min.css">
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.3.1/styles/github.min.css">
+    <link rel="stylesheet" href="css/main.css">
 
-    <style>
-      #button-container {
-        text-align: right; /* Alinea los botones a la derecha */
-      }
-    </style>
-  </head>
-  <body>
-    <div id="graph">
-      <h1>TC - PRACTICA 3 - GIC</h1>
-      <h2>Franco Rayas Angel Damian</h2>
+</head>
+<body>
+    <nav class="navbar navbar-expand-lg navbar color"> <!-- Barra de navegación en rojo -->
+        <a class="navbar-brand" href="./index.html">PDA Validador de sintaxis</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNavDropdown">
+            <ul class="navbar-nav">
+                <li class="nav-item active">
+                    <a class="nav-link" href="#">#Grafo <span class="sr-only">(current)</span></a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/three.html">Arbol</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">PDA</a>
+                </li>
+            </ul>
+        </div>
+    </nav>
+
+    <div class="center-vertically-and-horizontally">
+      <div class="text-center">
+        <h1>TC - Practica 3 - Gramaticas independientes de contexto</h1>
+        <h2>Franco Rayas Angel Damián</h2>
+        <input type="file" id="fileInput" class="btn color">
     </div>
-    <div id="container">
-      <div>
-        <p>Insertar aquí su archivo Java para su validación léxica.</p>
-        <label for="file-upload">Cargar archivo .java</label>
-        <input type="file" id="file-upload" accept=".java" />
+    </div>
+    <div class="center-vertically-and-horizontally">
+      <pre><code id="fileDisplayArea" class="language-c"></code></pre> <!-- Añade <code> y la clase para C -->
+  </div>
+  
+
+    <!-- Modal -->
+    <div class="modal fade" id="resultModal" tabindex="-1" role="dialog" aria-labelledby="resultModalLabel" aria-hidden="true">
+      <div class="modal-dialog modal-xl" role="document">
+        <div class="modal-content">
+          <div class="modal-header">
+            <h5 class="modal-title" id="resultModalLabel">Resultados</h5>
+            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+              <span aria-hidden="true">&times;</span>
+            </button>
+          </div>
+          <div class="modal-body">
+            <pre id="resultContent"></pre>
+          </div>
+          <div class="modal-footer">
+            <button type="button" class="btn btn-secondary" data-dismiss="modal">Cerrar</button>
+          </div>
+        </div>
       </div>
     </div>
 
-    <script>
-      document
-        .getElementById("file-upload")
-        .addEventListener("change", function (event) {
-          const file = event.target.files[0];
-          const reader = new FileReader();
-          reader.onload = function () {
-            const data = reader.result;
-            const url = new URL("preview.html", window.location.origin);
-            url.searchParams.append("file", file.name);
-            const previewWindow = window.open(url, "_blank");
-            previewWindow.onload = function () {
-              previewWindow.postMessage(data, window.location.origin);
-            };
-          };
-          reader.readAsText(file);
-        });
-    </script>
-  </body>
+    <!-- Bootstrap JS and dependencies -->
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/11.3.1/highlight.min.js"></script>
+    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.5.2/dist/umd/popper.min.js"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+    <script type="module" src="js/App.js"></script>
+</body>
 </html>
