+++ java/input.c
@@ -1,0 +1,18 @@
+A2 = A1 + 12 + C5 ;
+AB = A * B / 100 - 59 ;
+ABC = ( 340 % 2 ) + ( 12 - C ) ;
+AC = 0312 + ox12AB * ( 5.12E+12 + B ) ;
+Var1 = Var2 = Var3 = 8 ;
+VAR = ( CatA + ( ( CatA + CatB ) * CatC ) ) * ( CatD - CatF ) ;
+
+
+x = int * 2 + x ;
+info = cateto1 + cateto2 + ( ) - ) * ( ;
+VAR = ( CatA + ( CatA + CatB ) * CatC ) ) * ( CatD - CatF ) ;
+VAR = ( CatA + ( ( CatA + CatB ) * CatC ) ) * ( CatD - CatF ;
+temp = 0284 * 0x1X2 - 3.14EE1.3 ;
+
+typedef = val1 % val2 ;
+
+
+VAR = ( CatA + ( ( CatA + CatB ) * CatC ) ) * ( CatD - CatF ) ;
