--- java/input.java
@@ -1,6 +1,0 @@
-A2 = A1 + 12 + C5 ;
-AB = A * B / 100 - 59 ;
-ABC = ( 340 % 2 ) + ( 12 - C ) ;
-AC = 0312 + ox12AB * ( 5.12E+12 + B ) ;
-Var1 = Var2 = Var3 = 8 ;
-VAR = ( CatA + ( ( CatA + CatB ) * CatC ) ) * ( CatD - CatF ) ;
