+++ js/Stack.js
@@ -1,0 +1,29 @@
+// Stack.js
+export class Stack {
+  constructor() {
+    this.items = [];
+  }
+
+  push(element) {
+    this.items.push(element);
+  }
+
+  pop() {
+    if (this.items.length > 0) {
+      return this.items.pop();
+    }
+    return null;
+  }
+
+  top() {
+    return this.items[this.items.length - 1];
+  }
+
+  isEmpty() {
+    return this.items.length === 0;
+  }
+
+  clear() {
+    this.items = [];
+  }
+}
