+++ three.html
@@ -4,20 +4,22 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Árbol Dinámico D3.js</title>
+
     <style>
+
         body, html {
-          height: 100%;
-          margin: 0;
-          overflow: hidden;
-          font-family: Arial, sans-serif;
-          background-color: rgb(0, 0, 0);
-          color:#ccc;
+            height: 100%;
+            margin: 0;
+            overflow: hidden;
+            font-family: Arial, sans-serif;
+            background-color: #ffffff; /* Fondo blanco */
+            color: #EBB7A7; /* Texto en rosa claro */
         }
         .node circle {
-          fill: #fff;
-          color:#fff;
-          stroke-width: 3px;
-          opacity: 0.7;
+            fill: #EBB7A7; /* Rosa pastel */
+            stroke: #faa1bf; /* Borde en un tono de rosa un poco más oscuro */
+            stroke-width: 3px;
+            opacity: 0.7;
         }
 
         @keyframes animateDash {
@@ -30,21 +32,23 @@
             font: 12px sans-serif; 
         }
 
-        .link 
-        { 
-            fill: none; 
-            stroke: #ccc; 
-            stroke-width: 2px;  
+        .link {
+            fill: none;
+            stroke: #000000; /* Enlace en rosa muy claro */
+            stroke-width: 2px;
             opacity: 0.5;
             stroke-dasharray: 5, 5;
             stroke-linecap: round;
-            animation: animateDash 0.7s linear infinite; /* Aplica la animación infinita */
+            animation: animateDash 0.7s linear infinite;
         }
     </style>
+
 </head>
 <body>
+
+    
     <script src="https://d3js.org/d3.v6.min.js"></script>
     <script src="https://d3js.org/d3-drag.v2.min.js"></script>
-    <script src="script.js"></script> <!-- Asegúrate de enlazar correctamente tu script de JavaScript -->
+    <script src="./js/script.js"></script> <!-- Asegúrate de enlazar correctamente tu script de JavaScript -->
 </body>
 </html>
